finish actions refactor
